Needed to change Iout explicitly to a int64 integer

+++ vib_mod.f03
@@ -21,7 +21,7 @@
 !
       Contains
 
-      integer,parameter::IOut=6
+      integer(kind=int64),parameter::IOut=6
 
       
       end module vib_mod
